+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+
+## [0.12.0] - 2024-07-17
 ### Added
 - Request methods
 - PixReversal description